removed unused vector

+++ STEP_7/diag.f90
@@ -5,7 +5,7 @@
  use shared
  implicit none
 
- real(kind=8)    :: Ray, Gras, Four, Pr, beta_gras, &
+ real(kind=dp)    :: Ray, Gras, Four, Pr, beta_gras, &
                     Delta_Temp,vert_flux_t(0:Ny+1),horiz_flux_t(0:Nx+1), &
                     Nus, Fou, theta, Nus_comp, grad_T, Nus_comp_2, melt, &
                     melt_front, melt_front_expected, theta1, theta2
